ALL OF THIS NEEDS TESTING, CURRENTLY CRASHING THE APP.
run method to create markers for events, use addMarker(LatLng) pass the latitude and longitude.

when creating marker for event, call HashMap to get LatLng for building where event is, then pass to create marker

There is a time window in which +-1 hour, changable in globalVars, if in window event is added to HashMap and marked on Map, otherwise not considered.